itunes-backup  Alignment problems and failure on BE systems

+++ src/itunes_fmt_plug.c
@@ -52,7 +52,7 @@ john_register_one(&fmt_itunes);
 #define PLAINTEXT_LENGTH        125
 #define SALT_SIZE               sizeof(struct custom_salt)
 #define BINARY_ALIGN            1
-#define SALT_ALIGN              sizeof(int)
+#define SALT_ALIGN              sizeof(uint64_t)
 #ifdef SIMD_COEF_32
 #define MIN_KEYS_PER_CRYPT      (SSE_GROUP_SZ_SHA1 * SSE_GROUP_SZ_SHA256)
 #define MAX_KEYS_PER_CRYPT      (SSE_GROUP_SZ_SHA1 * SSE_GROUP_SZ_SHA256)
